chore: rename util
